# doc/api.yaml
openapi: 3.1.0
info:
  title: WASAPhoto
  description: |-
    Keep in touch with your friends by sharing photos of special moments, thanks to WASAPhoto!

    You can upload your photos directly from your PC, and they will be visible to everyone following you.
  version: 1.0.0
components:
  schemas:
    User:
      title: User
      description: This object represents a single user
      type: object
      properties:
        userId:
          type: integer
          description: Unique identifier for the user
          example: 1234
          readOnly: true
          minimum: 0
        username:
          description: The username displayed to the others
          type: string
          example: CoolUsername42
          minLength: 3
          maxLength: 32
          pattern: '^[A-Za-z0-9_\-]{3,32}$'
      required: [ "userId", "username" ]
    Profile:
      title: Profile
      description: This object represents the profile data of a user
      type: object
      properties:
        userId: { $ref: '#/components/schemas/User/properties/userId' }
        username: { $ref: '#/components/schemas/User/properties/username' }
        photoCount:
          description: The count of how many photos this user uploaded
          type: integer
          example: 42
          minimum: 0
        followersCount:
          description: The amount of users following this user
          type: integer
          example: 42
          minimum: 0
        followingCount:
          description: The amount of users followed by this user
          type: integer
          example: 42
          minimum: 0
        bannedCount:
          description: The amount of users banned by this user
          type: integer
          example: 42
          minimum: 0
      required: [ "userId", "username", "photoCount", "followersCount", "followingCount", "bannedCount" ]
...
    Photo:
      title: Photo
      description: This object represents a single photo in the user's feed
      type: object
      properties:
        photoId:
          type: integer
          description: Unique identifier for the photo
          example: 1234
          minimum: 0
          readOnly: true
        ownerId: { $ref: '#/components/schemas/User/properties/userId' }
        ownerUsername: { $ref: '#/components/schemas/User/properties/username' }
        image:
          type: string
          format: binary
          description: Binary data of the image
          minLength: 0
          maxLength: 999999999 # This is here to handle warnings, it's not a real limit
        mimeType:
          type: string
          description: The MIME type of the image
          example: "image/jpeg"
          pattern: '^[A-Za-z0-9_\-]{3,}$'
          minLength: 3
          maxLength: 32
        caption:
          type: string
          description: Caption for the image
          example: "A beautiful landscape"
          pattern: '^[\p{L}\p{N}\p{M}\p{P}\p{S} ]{0,32}$'
          minLength: 0
          maxLength: 32
        uploadTime:
          type: string
          format: date-time
          description: The date and time at which this photo was uploaded, RFC 3339 format
          example: 2017-07-21T17:32:28Z
        likeCount:
          type: integer
          description: Number of likes for the image
          example: 42
          minimum: 0
        commentsCount:
          type: integer
          description: Number of comments for the image
          example: 42
          minimum: 0
      required: [ "photoId", "ownerId", "image", "uploadTime", "likeCount", "commentsCount" ]
    Error:
      title: Error
      description: |-
        This object represents a single error to be returned for certain requests
      type: object
      properties:
        code:
          type: integer
          description: An HTTP error code
          example: 404
        message:
          type: string
          description: A brief message describing the error code
          example: The specified resource was not found
      required: [ "code", "message" ]
  parameters:
    usernameParam:
      name: username
      in: path
      description: Username of the user
      required: true
      content:
        application/json:
          schema: { $ref: '#/components/schemas/User/properties/username' }
...
  responses:
...
    BadRequest:
      description: The request was not compliant with the documentation (eg. missing fields, etc).
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    # 401
    Unauthorized:
      description: Invalid or expired authentication credentials.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    # 404
    NotFound:
      description: The specified resource was not found.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    # 409
    Conflict:
      description: The request could not be completed due to a conflict with the current state of the resource.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    # 422
    UnprocessableEntity:
      description: The request was well-formed but was unable to be followed due to semantic errors.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    # 500
    InternalServerError:
      description: The server encountered an internal error. Further info in server logs.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
security:
  - bearerAuth: [ ]
...
paths:
...
  /stream:
    get:
      tags: [ "User" ]
      operationId: getMyStream
      summary: Retrieve the content stream of the user
      description: |-
        Given a user's id, retrieve the content stream.

        The stream is composed of entries that have images, likes and comments.
        These entries are sorted in reverse chronological order.
      responses:
        "200":
          description: Successfully retrieved image stream for this user
          content:
            application/json:
              schema:
                type: object
                description: photo stream schema
                properties:
                  photos:
                    type: array
                    description: The list of photos this user uploaded
                    items: { $ref: '#/components/schemas/Photo' }
                    uniqueItems: true
                    minItems: 0
                    maxItems: 99999
        "400": { $ref: '#/components/responses/BadRequest' }
        "401": { $ref: '#/components/responses/Unauthorized' }
        "500": { $ref: '#/components/responses/InternalServerError' }
  /users:
    get:
      tags: [ "User" ]
      operationId: getAllUsers
      summary: Retrieve all users
      description: Fetch all users in the database via paginated requests.
      parameters:
        - name: page
          in: query
          description: Page number
          required: false
          schema:
            type: integer
            minimum: 1
            default: 1
        - name: pageSize
          in: query
          description: Number of items per page
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 50
      responses:
        "200":
          description: Users retrieved successfully
          content:
            application/json:
              schema:
                type: array
                description: list of users
                items: { $ref: '#/components/schemas/User' }
                uniqueItems: true
                minItems: 0
                maxItems: 99999
        "400": { $ref: '#/components/responses/BadRequest' }
        "500": { $ref: '#/components/responses/InternalServerError' }
  /users/{username}:
    parameters:
      - $ref: '#/components/parameters/usernameParam'
    put:
      tags: [ "User" ]
      operationId: setMyUserName
      summary: Change your username
      description: |-
        Given a user's username, update said username.
      requestBody:
        description: The new username to set
        required: true
        content:
          application/json:
            schema:
              type: object
              description: username schema
              properties:
                username: { $ref: '#/components/schemas/User/properties/username' }
      responses:
        "201":
          description: Username changed successfully
          content:
            application/json:
              schema: { $ref: '#/components/schemas/User' }
        "400": { $ref: '#/components/responses/BadRequest' }
        "401": { $ref: '#/components/responses/Unauthorized' }
        "404": { $ref: '#/components/responses/NotFound' }
        "409": { $ref: '#/components/responses/Conflict' }
        "422": { $ref: '#/components/responses/UnprocessableEntity' }
        "500": { $ref: '#/components/responses/InternalServerError' }
  /users/{username}/profile:
    parameters:
      - $ref: '#/components/parameters/usernameParam'
    get:
      tags: [ "User" ]
      operationId: getUserProfile
      summary: Retrieve the information of a user
      description: |-
        Given a user's username, retrieve all the public info available.
      responses:
        "200":
          description: Successfully retrieved profile data
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Profile' }
        "400": { $ref: '#/components/responses/BadRequest' }
        "404": { $ref: '#/components/responses/NotFound' }
        "422": { $ref: '#/components/responses/UnprocessableEntity' }
        "500": { $ref: '#/components/responses/InternalServerError' }
...
  /users/{username}/bans/count:
    parameters:
      - $ref: '#/components/parameters/usernameParam'
    get:
      tags: [ "Ban" ]
      operationId: getBansCount
      summary: Get the count of users banned by the current user
      description: Get the count of users banned by the current user
      responses:
        "200":
          description: Successfully retrieved the count of banned users
          content:
            application/json:
              schema:
                type: integer
                description: count of banned users
                example: 42
                minimum: 0
        "400": { $ref: '#/components/responses/BadRequest' }
        "401": { $ref: '#/components/responses/Unauthorized' }
        "404": { $ref: '#/components/responses/NotFound' }
        "500": { $ref: '#/components/responses/InternalServerError' }
...
  /users/{username}/following/count:
    parameters:
      - $ref: '#/components/parameters/usernameParam'
    get:
      tags: [ "Follow" ]
      operationId: getFollowingCount
      summary: Get the count of users followed by the current user
      description: Get the count of users followed by the current user
      responses:
        "200":
          description: Successfully retrieved the count of followed users
          content:
            application/json:
              schema:
                type: integer
                description: count of followed users
                example: 42
                minimum: 0
        "400": { $ref: '#/components/responses/BadRequest' }
        "401": { $ref: '#/components/responses/Unauthorized' }
        "404": { $ref: '#/components/responses/NotFound' }
        "422": { $ref: '#/components/responses/UnprocessableEntity' }
        "500": { $ref: '#/components/responses/InternalServerError' }
...
  /users/{username}/followers/count:
    parameters:
      - $ref: '#/components/parameters/usernameParam'
    get:
      tags: [ "Follow" ]
      operationId: getFollowersCount
      summary: Get the count of users following the current user
      description: Get the count of users following the current user
      responses:
        "200":
          description: Successfully retrieved the count of followers
          content:
            application/json:
              schema:
                type: integer
                description: count of followers
                example: 42
                minimum: 0
        "400": { $ref: '#/components/responses/BadRequest' }
        "401": { $ref: '#/components/responses/Unauthorized' }
        "404": { $ref: '#/components/responses/NotFound' }
        "422": { $ref: '#/components/responses/UnprocessableEntity' }
        "500": { $ref: '#/components/responses/InternalServerError' }
...